+++ CHANGELOG.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Utilitário `convert_name_to_uf`
-- Utilitário `is_valid_natureza_juridica` [#641](https://github.com/brazilian-utils/python/issues/641)
-- Utilitário `get_natureza_juridica_description` [#641](https://github.com/brazilian-utils/python/issues/641)
-- Utilitário `list_all_natureza_juridica` [#641](https://github.com/brazilian-utils/python/issues/641)
+- Utilitário `is_valid_legal_nature` [#641](https://github.com/brazilian-utils/python/issues/641)
+- Utilitário `get_legal_nature_description` [#641](https://github.com/brazilian-utils/python/issues/641)
+- Utilitário `list_all_legal_nature` [#641](https://github.com/brazilian-utils/python/issues/641)
 
 ## [2.3.0] - 2025-10-07